refactor(rowbinary): drop the _writer suffix from src/writers files
The writers/ directory already disambiguates from readers/, so the suffix was
redundant: src/writers/integers_writer.ts -> src/writers/integers.ts, etc.
(now mirroring the reader basenames). Updated the writer barrel, sibling
imports, test imports, and the writer.md reference table; subpath exports become
@clickhouse/rowbinary/writers/<type>.
