+++ skills/clickhouse-js-node-rowbinary/src/writers/aggregateFunction.ts
@@ -1,4 +1,4 @@
-import { type Writer } from "./core_writer.js";
+import { type Writer } from "./core.js";
 
 /**
  * Inverse of `readAggregateFunction`: an `AggregateFunction(func, T…)` column is
