+++ skills/clickhouse-js-node-rowbinary/src/writers/bool.ts
@@ -1,5 +1,5 @@
-import { Sink } from "./core_writer.js";
-import { writeUInt8 } from "./integers_writer.js";
+import { Sink } from "./core.js";
+import { writeUInt8 } from "./integers.js";
 
 /**
  * Write a `Bool`: 1 byte, stored as `UInt8` (`false` -> 0, `true` -> 1). Mirror
