+++ skills/clickhouse-js-node-rowbinary/src/writers/datetime.ts
@@ -1,4 +1,4 @@
-import { type Writer, Sink, reserve } from "./core_writer.js";
+import { type Writer, Sink, reserve } from "./core.js";
 import { type Microseconds, type Nanoseconds } from "../readers/datetime.js";
 
 const MS_PER_DAY = 86_400_000;
