+++ skills/clickhouse-js-node-rowbinary/tests/AggregateFunction.write.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 import { encode } from "./encode.js";
-import { writeAggregateFunction } from "../src/writers/aggregateFunction_writer.js";
+import { writeAggregateFunction } from "../src/writers/aggregateFunction.js";
 
 describe("writeAggregateFunction", () => {
   it("throws — opaque, unframed state is not generically encodable", () =>
